app/requirements.txt
Django==2.2.14
psycopg2-binary==2.8.5
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
Markdown==3.2.2
django-filter==2.3.0

